split fp8/int8 tests
