+++ test/srt/run_suite.py
@@ -312,7 +312,8 @@ suite_xeon = {
         TestFile("cpu/test_topk.py"),
         TestFile("test_cpu_graph.py"),
         TestFile("test_intel_amx_attention_backend.py"),
-        TestFile("test_intel_amx_attention_backend_quant.py"),
+        TestFile("test_intel_amx_attention_backend_1.py"),
+        TestFile("test_intel_amx_attention_backend_2.py"),
     ],
 }
