+++ test/srt/test_intel_amx_attention_backend.py
@@ -1,6 +1,6 @@
 """
 Usage:
-python3 -m unittest test_intel_amx_attention_backend.TestIntelAMXAttnBackend.test_mmlu
+python3 -m unittest test_intel_amx_attention_backend.TestIntelAMXAttnBackend.test_latency_default_model
 """
 
 import unittest
