+++ test/srt/test_intel_amx_attention_backend_2.py
@@ -1,13 +1,12 @@
 """
+For intel_amx attention backend w8a8 tests
 Usage:
-python3 -m unittest test_intel_amx_attention_backend_quant.TestIntelAMXAttnBackendQuant.test_latency_fp8_qwen
+python3 -m unittest test_intel_amx_attention_backend_2.TestIntelAMXAttnBackendQuant.test_latency_w8a8_default_model
 """
 
 import unittest
 
 from sglang.test.test_utils import (
-    DEFAULT_MODEL_NAME_FOR_TEST_FP8_WITH_MOE,
-    DEFAULT_MODEL_NAME_FOR_TEST_QWEN_FP8,
     DEFAULT_MODEL_NAME_FOR_TEST_W8A8,
     DEFAULT_MODEL_NAME_FOR_TEST_W8A8_WITH_MOE,
     CustomTestCase,
@@ -16,14 +15,6 @@ from sglang.test.test_utils import (
 
 
 class TestIntelAMXAttnBackendQuant(CustomTestCase):
-
-    @intel_amx_benchmark(extra_args=["--batch-size", "4"], min_throughput=150)
-    def test_latency_fp8_qwen(self):
-        return DEFAULT_MODEL_NAME_FOR_TEST_QWEN_FP8
-
-    @intel_amx_benchmark(extra_args=["--batch-size", "4"], min_throughput=50)
-    def test_latency_fp8_moe_model(self):
-        return DEFAULT_MODEL_NAME_FOR_TEST_FP8_WITH_MOE
 
     @intel_amx_benchmark(
         extra_args=["--batch-size", "4", "--quantization", "w8a8_int8"],
